:recycle:[Refactor] searchProductList 함수 유틸함수로 추가 #171
